change query snippet detail api, use handler snippetConfig, some config not have batch-query-snippet-yml api
