+++ modules/pipeline/services/pipelinesvc/action_output_handler_test.go
@@ -11,7 +11,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-package snippetsvc
+package pipelinesvc
 
 import (
 	"testing"
